location stacks: computing and drawing (#24)
* test drawing/discretization: Masado's example

* discretization: location stacks

* drawing: location stack, aligned, colored
---------

+++ src/dyada/coordinates.py
@@ -39,7 +39,7 @@ Coordinate: TypeAlias = npt.NDArray[np.float64]
 
 
 def coordinate_from_sequence(c: Sequence[float]) -> Coordinate:
-    return np.asarray(c, dtype=np.float32)
+    return np.asarray(c, dtype=np.float64)
 
 
 class CoordinateInterval(NamedTuple):
